+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+20180409
+
+- Add Gradient Filter.
+
 20180328
 
 - Update README